little clean up

+++ WebCrawler/MongoAccess.py
@@ -12,6 +12,7 @@ class MongoAccess:
 
     def __init__(self, collection_name):
         self.__collection_name = collection_name
+        self.__mongod_process = None
 
     def start_mongo_server(self):
         start_command = self.__MONGO_BIN + " --dbpath " + self.__DB_PATH
@@ -20,7 +21,9 @@ class MongoAccess:
         self.collection = self.client["lens_db." + self.__collection_name]
 
     def stop_mongo_server():
-        self.__mongod_process.terminate()
+        if(self.__mongod_process != None):
+            self.__mongod_process.terminate()
+            self.__mongod_process = None
 
     def add_lens(self, lens):
         self.collection.insert_one(lens)
@@ -28,10 +31,6 @@ class MongoAccess:
     def update_lens(self, lens):
         lensname = lens[self.__KEY_LENSNAME]
         return self.collection.update_one({self.__KEY_LENSNAME: lensname}, {'$set': lens}) 
-
-    def find_lens(self, lens):
-        lensname = lens[self.__KEY_LENSNAME]
-        return self.collection.find({self.__KEY_LENSNAME: lensname})
 
     def find_all_lenses(self):
         return self.collection.find({})
